Create notes package, move files and update references
