+++ src/main/java/edu/stanford/bmir/protege/web/client/ui/portlet/bioportal/notes/BioPortalProposalsPortlet.java
@@ -1,4 +1,4 @@
-package edu.stanford.bmir.protege.web.client.ui.portlet.bioportal;
+package edu.stanford.bmir.protege.web.client.ui.portlet.bioportal.notes;
 
 import com.google.gwt.core.client.GWT;
 import com.google.gwt.event.dom.client.ClickEvent;
