+++ src/main/java/edu/stanford/bmir/protege/web/client/ui/portlet/bioportal/notes/NewNotePanel.java
@@ -1,4 +1,4 @@
-package edu.stanford.bmir.protege.web.client.ui.portlet.bioportal;
+package edu.stanford.bmir.protege.web.client.ui.portlet.bioportal.notes;
 
 import com.gwtext.client.core.EventObject;
 import com.gwtext.client.core.Position;
